requirements.txt
alembic==1.10.2
async-generator==1.10
attrs==23.1.0
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
dnspython==2.3.0
email-validator==1.3.1
exceptiongroup==1.1.1
Flask==2.2.3
Flask-Login==0.6.2
Flask-Migrate==4.0.4
Flask-SQLAlchemy==3.0.3
Flask-Testing==0.8.1
Flask-WTF==1.1.1
greenlet==2.0.2
h11==0.14.0
idna==3.4
importlib-metadata==6.1.0
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.4
MarkupSafe==2.1.2
outcome==1.2.0
packaging==23.1
pluggy==1.0.0
pydantic==1.10.7
PySocks==1.7.1
pytest==7.3.0
python-dotenv==1.0.0
requests==2.29.0
selenium==4.9.1
sniffio==1.3.0
sortedcontainers==2.4.0
SQLAlchemy==2.0.7
tomli==2.0.1
trio==0.22.0
trio-websocket==0.10.2
typing_extensions==4.5.0
urllib3==1.26.15
Werkzeug==2.2.3
wsproto==1.2.0
WTForms==3.0.1
zipp==3.15.0
